+++ examples/alpine/options.ts
@@ -1,5 +1,0 @@
-export default {
-    features: {
-        alpine: true,
-    },
-};