+++ examples/docs/remedy.config.ts
@@ -2,6 +2,5 @@ export default {
     features: {
         fontawesome: true,
         tailwind: true,
-        alpine: false,
     },
 };