+++ packages/server/server/options.ts
@@ -12,6 +12,7 @@ interface ServerOptionsFeatures {
     alpine?: boolean;
     htmx?: HtmxOptions;
     sse?: boolean;
+    typescript?: boolean;
     static?: boolean;
     dev?: boolean;
 }
@@ -27,6 +28,7 @@ const options = {
             debug: false,
         },
         sse: false,
+        typescript: true,
         static: true,
         dev: import.meta.env.NODE_ENV === "development",
     },